Imporving error handlidng, data validation, adding time stapms

+++ graphQL/resolvers.js
@@ -1,43 +1,66 @@
 const User = require('../models/User');
 const Product = require('../models/Product');
 const jwt = require('jsonwebtoken');
+const bcrypt = require('bcryptjs');
+const validator = require('validator');
+const jwt = require('jsonwebtoken');
 
 module.exports = {
 
   createUser: async function({ userInput }, req) {
 
-        //   const email = args.userInput.email;
-        const existingUser = await User.findOne({ email: userInput.email });
-        if (existingUser) {
-          const error = new Error('User exists already!');
-          throw error;
-        }
-        const hashedPw = userInput.password;
-        const user = new User({
-          email: userInput.email,
-          password: hashedPw
-        });
-        const createdUser = await user.save();
-        return { ...createdUser._doc, _id: createdUser._id.toString() };
+    const errors = [];
+    if (validator.isEmpty(userInput.email) || 
+        validator.isEmpty(userInput.password) || 
+        validator.isEmpty(userInput.name) ){
+      errors.push({ message: 'Required fields empty' });
+    }
+    if (!validator.isEmail(userInput.email)) {
+      errors.push({ message: 'E-Mail is invalid.' });
+    }
+    if (errors.length > 0) {
+      const error = new Error('Invalid input.');
+      error.data = errors;
+      error.code = 422;
+      throw error;
+    }
+    const existingUser = await User.findOne({ email: userInput.email });
+    if (existingUser) {
+      const error = new Error('User exists already!');
+      error.code = 422;
+      throw error;
+    }
+    const hashedPw = await bcrypt.hash(userInput.password, 12);
+    const user = new User({
+      email: userInput.email,
+      password: hashedPw,
+      name: userInput.name
+    });
+    const createdUser = await user.save();
+    const token = jwt.sign({userId: createdUser._id.toString()}, 'secret', {expiresIn: '24h'});    
+    return { ...createdUser._doc, _id: createdUser._id.toString(), token };
   },
 
   login: async function({ email, password }) {
 
-    const user = await User.findOne({email:email, password:password});
-
+    if (validator.isEmpty(email) || validator.isEmpty(password)) {
+      const error = new Error('Required fields empty');
+      error.code = 422;
+      throw error;
+    }
+    const user = await User.findOne({email:email});
     if (!user) {
-      const error = new Error('Incorrect user or password');
+      const error = new Error('User not found');
       error.code = 401;
       throw error;
     }
-
-    const token = jwt.sign(
-      {
-        email: user.email
-      },
-      'secret',
-      { expiresIn: '1h' }
-    );
+    const isEqual = await bcrypt.compare(password, user.password);
+    if (!isEqual) {
+      const error = new Error('Password is incorrect.');
+      error.code = 401;
+      throw error;
+    }
+    const token = jwt.sign({userId: user._id.toString()}, 'secret', {expiresIn: '24h'});
     return { token: token };
   },
 
@@ -51,7 +74,16 @@ module.exports = {
 
       throw error;
     }
-
+    if(validator.isEmpty(productInput.name) || 
+        validator.isEmpty(productInput.price) ||
+        validator.isEmpty(productInput.description) ||
+        validator.isEmpty(productInput.ammount) ||
+        validator.isEmpty(productInput.imgURL)
+      ){
+      const error = new Error('Required fields empty');
+      error.code = 422;
+      throw error;
+    }
     const product = new Product({
       name:  productInput.name,
       price: productInput.price,
@@ -62,10 +94,12 @@ module.exports = {
 
     const createdProduct =  await product.save();
 
-    console.log(createdProduct);
-
-    return true;
-
+    return {
+      ...createdProduct._doc,
+      _id: createdProduct._id.toString(),
+      createdAt: createdProduct.createdAt.toISOString(),
+      updatedAt: createdProduct.updatedAt.toISOString()
+    };;
   },
 
   updateProduct: async function({productInput}, req){
@@ -79,6 +113,17 @@ module.exports = {
       throw err;
     }
 
+    if(validator.isEmpty(productInput.name) || 
+     validator.isEmpty(productInput.price) ||
+     validator.isEmpty(productInput.description) ||
+     validator.isEmpty(productInput.ammount) ||
+     validator.isEmpty(productInput.imgURL)
+    ){
+     const error = new Error('Required fields empty');
+     error.code = 422;
+     throw error;
+    }
+
     const product = await Product.findById(productInput.id);
 
     product.name = productInput.name;
@@ -87,9 +132,14 @@ module.exports = {
     product.ammount = productInput.ammount;
     product.imgURL = productInput.imgURL;
 
-    await product.save();
+    const updatedProduct = await product.save();
 
-    return true;
+    return {
+      ...updatedProduct._doc,
+      _id: updatedProduct._id.toString(),
+      createdAt: updatedProduct.createdAt.toISOString(),
+      updatedAt: updatedProduct.updatedAt.toISOString()
+    };
   },
 
   deleteProduct: async function({id}, req){
@@ -101,6 +151,12 @@ module.exports = {
       err.code = 401;
 
       throw err;
+    }
+
+    if(validator.isEmpty(productInput.name)){
+      const error = new Error('Required fields empty');
+     error.code = 422;
+     throw error;
     }
 
     await Product.findByIdAndRemove(id);
@@ -118,7 +174,12 @@ module.exports = {
 
       throw err;
     }
-  
+
+    if(validator.isEmpty(offset) || validator.isEmpty(limit)){
+     const error = new Error('Required fields empty');
+     error.code = 422;
+     throw error;
+    }
     const products = await Product.find().skip(offset).limit(limit);
 
     return {products: products};
@@ -135,10 +196,12 @@ module.exports = {
       throw err;
     }
 
+    if(validator.isEmpty(id)){
+      const error = new Error('Required fields empty');
+      error.code = 422;
+      throw error;
+     }
     const product = await Product.findById(id);
-
     return product;
-
   }
-
 }
